requirements.txt
faiss-cpu==1.7.4
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
numpy==1.26.4
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
python-magic==0.4.27
streamlit==1.31.1
tiktoken==0.6.0
unstructured==0.12.4